Tue, Apr 23, 2024, 2:24 PM -03:00

+++ filter-less/helpers.c
@@ -10,7 +10,7 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
         for (int j = 0; j < width ; j++)
         {
             // Average of RGB values in all pixels
-             int avg_rgb = (image[i][j].rgbtRed + image[i][j].rgbtGreen + image[i][j].rgbtBlue) / 3.0;
+             float avg_rgb = (image[i][j].rgbtRed + image[i][j].rgbtGreen + image[i][j].rgbtBlue) / 3.0;
              avg_rgb = round(avg_rgb);
 
              // Update each pixel of the image
